Made initial structure
