--- Jantimizer/Histograms/HistogramBucket.cs
@@ -1,14 +1,0 @@
-﻿namespace Histograms
-{
-    public class HistogramBucket : IHistogramBucket
-    {
-        public int ValueStart { get; }
-        public int Count { get; set; }
-
-        public HistogramBucket(int valueStart, int count)
-        {
-            ValueStart = valueStart;
-            Count = count;
-        }
-    }
-}
