+++ Jantimizer/Histograms/HistogramModels/HistogramBucket.cs
@@ -1,0 +1,14 @@
+﻿namespace Histograms
+{
+    public class HistogramBucket : IHistogramBucket
+    {
+        public int ValueStart { get; }
+        public int Count { get; set; }
+
+        public HistogramBucket(int valueStart, int count)
+        {
+            ValueStart = valueStart;
+            Count = count;
+        }
+    }
+}
