+++ Jantimizer/Histograms/HistogramModels/IHistogram.cs
@@ -1,0 +1,14 @@
+﻿using System.Data;
+
+namespace Histograms
+{
+    public interface IHistogram
+    {
+        public string TableName { get; }
+        public string AttributeName { get; }
+        public List<IHistogramBucket> Buckets { get; }
+
+        public void GenerateHistogram(DataTable table, string key);
+        public void GenerateHistogram(List<int> column);
+    }
+}
