+++ Jantimizer/Histograms/HistogramModels/IHistogramBucket.cs
@@ -1,0 +1,8 @@
+﻿namespace Histograms
+{
+    public interface IHistogramBucket
+    {
+        public int ValueStart { get; }
+        public int Count { get; set; }
+    }
+}
