--- Jantimizer/Histograms/IHistogram.cs
@@ -1,12 +1,0 @@
-﻿using System.Data;
-
-namespace Histograms
-{
-    public interface IHistogram
-    {
-        public List<IHistogramBucket> Buckets { get; }
-
-        public void GenerateHistogram(DataTable table, string key);
-        public void GenerateHistogram(List<int> column);
-    }
-}
