--- Jantimizer/Histograms/IHistogramBucket.cs
@@ -1,8 +1,0 @@
-﻿namespace Histograms
-{
-    public interface IHistogramBucket
-    {
-        public int ValueStart { get; }
-        public int Count { get; set; }
-    }
-}
